+++ blocks/news-cards/news-cards.css
@@ -1,0 +1,85 @@
+/* Cards container */
+.news-cards {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 20px;
+  align-items: stretch;
+  width: 100%;
+  margin-top: 20px;
+}
+
+/* Card grid layout */
+.news-cards > ul {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(257px, 1fr));
+  grid-gap: 24px;
+}
+
+/* Individual card styles */
+.news-card {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  background-color: #eaeaea;
+  border-bottom: 2px solid #02084b;
+  text-decoration: none;
+  letter-spacing: .01em;
+}
+
+.news-card-content {
+  padding: 20px 30px;
+}
+
+.news-card-title {
+  margin: 0 0 10px 0;
+  font-family: "Zilla Slab";
+  color: black;
+  text-transform: none;
+  line-height: 1.2;
+  font-size: 1.375rem;
+  font-weight: 500;
+  text-align: left;
+}
+
+.news-card-description {
+  font-size: 16px;
+  color: #000;
+  margin-bottom: 10px;
+  font: 400 16px / 1.5 'Open Sans', sans-serif;
+  text-align: left;
+}
+
+.news-card-image {
+  width: 100%;
+  max-width: 100%;
+  height: 182px;
+  object-fit: cover;
+}
+
+.news-card-category,
+.news-card-date {
+  font-size: 14px;
+  color: #000;
+  margin-bottom: 10px;
+  text-align: left;
+}
+
+/* State preservation */
+.news-card,
+.news-card:hover,
+.news-card:focus,
+.news-card:active,
+.news-card:visited,
+.news-card *,
+.news-card *:hover,
+.news-card *:focus,
+.news-card *:active {
+  color: inherit !important;
+  text-decoration: none !important;
+}
