+++ templates/news-landing/news-landing.js
@@ -1,363 +1,18 @@
-import { getMetadata, buildBlock, loadBlock, decorateBlock } from '../../scripts/aem.js';
-import { div, h1 } from '../../scripts/dom-helpers.js';
+import { buildBlock, loadBlock, decorateBlock } from '../../scripts/aem.js';
 
-// Constants
-const ITEMS_PER_PAGE = 6;
-const DEFAULT_IMAGE = "/for-parents/media_14d486ea46814365c71f7d6d83b417e682cf318f9.jpeg";
-const DESCRIPTION_MAX_LENGTH = 138;
 
-/**
- * Extracts filter value from current URL
- * @returns {string} Current filter value
- */
-function getFilterFromURL() {
-  const url = new URL(window.location.href);
-  
-  // Check for query parameter first
-  const filterParam = url.searchParams.get('filterParam');
-  if (filterParam) {
-    // Map encoded text back to filter value
-    const filterMap = {
-      'the past 7 days': '7days',
-      'the past 30 days': '30days',
-      'the past 90 days': '90days',
-      'all time': 'all'
-    };
-    return filterMap[decodeURIComponent(filterParam)] || 'all';
-  }
-  
-  // Check for year in path (e.g., /news/2015.html)
-  const pathMatch = url.pathname.match(/\/news\/(\d{4})\.html$/);
-  if (pathMatch) return pathMatch[1];
-  
-  return 'all'; // Default value
-}
-
-/**
- * Updates the URL based on selected filter
- * @param {string} filterValue - Selected filter value
- */
-function updateURL(filterValue) {
-  const baseUrl = '/news';
-  
-  // Map filter values to their display text
-  const filterTextMap = {
-    '7days': 'the past 7 days',
-    '30days': 'the past 30 days',
-    '90days': 'the past 90 days',
-    'all': 'all time'
-  };
-
-  if (Object.keys(filterTextMap).includes(filterValue)) {
-    const filterText = filterTextMap[filterValue];
-    const newUrl = `${baseUrl}.html?filterParam=${encodeURIComponent(filterText)}`;
-    window.history.pushState({ path: newUrl }, '', newUrl);
-  } else {
-    // Year filter remains the same
-    const newUrl = `${baseUrl}/${filterValue}.html`;
-    window.history.pushState({ path: newUrl }, '', newUrl);
-  }
-}
-
-/**
- * Creates the filter section with dropdown and button
- * @param {Function} onFilterChange - Callback for filter changes
- * @returns {HTMLElement} Filter container element
- */
-function createFilterSection(onFilterChange) {
-  const filterContainer = document.createElement('div');
-  filterContainer.className = 'news-filter-container';
-
-  const filterElements = {
-    label: createFilterLabel(),
-    select: createFilterDropdown(),
-    button: createFilterButton(onFilterChange)
-  };
-
-  filterContainer.append(filterElements.label, filterElements.select, filterElements.button);
-  return filterContainer;
-}
-
-/**
- * Creates the filter label element
- * @returns {HTMLElement} Label element
- */
-function createFilterLabel() {
-  const label = document.createElement('span');
-  label.className = 'sws-advance-search__text';
-  label.setAttribute('aria-hidden', 'true');
-  label.textContent = 'Show me all news from';
-  return label;
-}
-
-/**
- * Creates the filter dropdown with time period options
- * @returns {HTMLElement} Select element
- */
-function createFilterDropdown() {
-  const select = document.createElement('select');
-  select.id = 'filterOption';
-  select.className = 'sws-advance-search__select';
-  select.setAttribute('aria-label', 'Show me news for');
-
-  const timeOptions = [
-    { value: 'all', text: 'all time' },
-    { value: '7days', text: 'the past 7 days' },
-    { value: '30days', text: 'the past 30 days' },
-    { value: '90days', text: 'the past 90 days' }
-  ];
-
-  const yearOptions = Array.from({ length: 14 }, (_, i) => ({
-    value: String(2024 - i),
-    text: String(2024 - i)
-  }));
-
-  const currentFilter = getFilterFromURL();
-
-  [...timeOptions, ...yearOptions].forEach(opt => {
-    const option = document.createElement('option');
-    option.value = opt.value;
-    option.textContent = opt.text;
-    option.selected = opt.value === currentFilter;
-    select.appendChild(option);
-  });
-
-  return select;
-}
-
-/**
- * Creates the filter action button
- * @param {Function} onFilterChange - Callback when filter is applied
- * @returns {HTMLElement} Button element
- */
-function createFilterButton(onFilterChange) {
-  const button = document.createElement('button');
-  button.className = 'news-filter-button';
-  button.textContent = 'Go';
-  
-  button.addEventListener('click', () => {
-    const select = document.querySelector('#filterOption');
-    const selectedValue = select.value;
-    updateURL(selectedValue);
-    onFilterChange(selectedValue);
-  });
-  
-  return button;
-}
-
-/**
- * Creates loading indicator with animation
- * @returns {HTMLElement} Loading indicator element
- */
-function createLoadingIndicator() {
-  const indicator = document.createElement('div');
-  indicator.className = 'news-view-more';
-  indicator.innerHTML = `
-    <div class="sws-loader-svg-1">
-      <svg width="40" height="40" viewBox="0 0 44 44" xmlns="http://www.w3.org/2000/svg" stroke="#000">
-        <g fill="none" fill-rule="evenodd" stroke-width="2">
-          <circle cx="22" cy="22" r="1">
-            <animate attributeName="r" begin="0s" dur="1.8s" values="1; 20" calcMode="spline" keyTimes="0; 1" keySplines="0.165, 0.84, 0.44, 1" repeatCount="indefinite"/>
-            <animate attributeName="stroke-opacity" begin="0s" dur="1.8s" values="1; 0" calcMode="spline" keyTimes="0; 1" keySplines="0.3, 0.61, 0.355, 1" repeatCount="indefinite"/>
-          </circle>
-          <circle cx="22" cy="22" r="1">
-            <animate attributeName="r" begin="-0.9s" dur="1.8s" values="1; 20" calcMode="spline" keyTimes="0; 1" keySplines="0.165, 0.84, 0.44, 1" repeatCount="indefinite"/>
-            <animate attributeName="stroke-opacity" begin="-0.9s" dur="1.8s" values="1; 0" calcMode="spline" keyTimes="0; 1" keySplines="0.3, 0.61, 0.355, 1" repeatCount="indefinite"/>
-          </circle>
-        </g>
-      </svg>
-    </div>`;
-  indicator.style.display = 'none';
-  return indicator;
-}
-
-/**
- * Creates news card HTML structure
- * @param {Object} item - News item data
- * @returns {string} HTML string for news card
- */
-function createNewsCardHTML(item) {
-  return `
-    <a href="${item.path}" class="news-card">
-      <img class="news-card-image" src="${item.image || DEFAULT_IMAGE}" alt="${item.title}"/>
-      <div class="news-card-content">
-        <div class="news-card-category">/ News category</div>
-        <h3 class="news-card-title">${item.title || ''}</h3>
-        <div class="news-card-date">21 Oct 2024</div>
-        <div class="news-card-description">${getTrimmedDescription(item.description || '')}</div>
-      </div>
-    </a>`;
-}
-
-/**
- * Trims description text to specified length
- * @param {string} description - Original description text
- * @returns {string} Trimmed description
- */
-function getTrimmedDescription(description) {
-  if (description.length > DESCRIPTION_MAX_LENGTH) {
-    return description.substring(0, DESCRIPTION_MAX_LENGTH - 3) + '...';
-  }
-  return description;
-}
-
-/**
- * Filters news items based on selected filter
- * @param {Array} items - All news items
- * @param {string} filterValue - Selected filter value
- * @returns {Array} Filtered news items
- */
-function filterNewsItems(items, filterValue) {
-  if (!items.length) return [];
-
-  switch (filterValue) {
-    case '7days':
-    case '30days':
-    case '90days': {
-      const days = parseInt(filterValue);
-      const cutoffDate = new Date();
-      cutoffDate.setDate(cutoffDate.getDate() - days);
-      return items.filter(item => new Date(item.date) >= cutoffDate);
-    }
-    case 'all':
-      return items;
-    default:
-      // Year filter
-      return items.filter(item => item.path.includes(`/news/${filterValue}`));
-  }
-}
-
-/**
- * Fetches and formats news items from API
- * @param {string} filterValue - Current filter value
- * @returns {Promise<Array>} Formatted news items
- */
-async function fetchNewsItems(filterValue = 'all') {
-  try {
-    const response = await fetch('/query-index.json');
-    const { data } = await response.json();
-    
-    const filteredData = filterNewsItems(data, filterValue);
-    return filteredData.map(item => [createNewsCardHTML(item)]);
-  } catch (error) {
-    console.error('Error fetching news items:', error);
-    return [];
-  }
-}
-
-/**
- * Creates no results message container
- * @returns {HTMLElement} No results container
- */
-function createNoResultsMessage() {
-  const noResultsContainer = document.createElement('div');
-  noResultsContainer.className = 'news-no-results-container';
-  
-  const noResultsMessage = document.createElement('div');
-  noResultsMessage.className = 'news-no-results-message';
-  noResultsMessage.textContent = "We didn't find anything that matches your filters. Try expanding your search or browsing the whole collection.";
-  
-  noResultsContainer.appendChild(noResultsMessage);
-  return noResultsContainer;
-}
-
-/**
- * Creates horizontal divider
- * @returns {HTMLElement} Divider element
- */
-function createDivider() {
-  const container = document.createElement('div');
-  container.className = 'news-content-divider-container';
-  
-  const line = document.createElement('div');
-  line.className = 'news-content-divider-line';
-  
-  container.appendChild(line);
-  return container;
-}
-/**
- * Main block decoration function
- * @param {HTMLElement} block - Block element to decorate
- */
 export default async function decorate(block) {
   try {
-    // Initialize containers
-    const contentSection = document.createElement('div');
-    contentSection.className = 'news-content-section';
+    // Create news section block which will handle both filter and content
+    const newsSection = buildBlock('news-section', []);
     
-    const newsContentContainer = document.createElement('div');
-    newsContentContainer.className = 'news-content-container';
+    // Add section to block
+    block.append(newsSection);
+    decorateBlock(newsSection);
     
-    // Initialize state
-    let currentPage = 1;
-    let allItems = [];
-    const loadingIndicator = createLoadingIndicator();
-
-    /**
-     * Updates displayed news items
-     * @param {number} page - Page number to display
-     */
-    const showItems = async (page) => {
-      const existingButton = newsContentContainer.querySelector('.news-more-button');
-      if (existingButton) existingButton.style.display = 'none';
-      loadingIndicator.style.display = 'block';
-
-      await new Promise(resolve => setTimeout(resolve, 800));
-
-      newsContentContainer.textContent = '';
-      
-      if (allItems.length === 0) {
-        newsContentContainer.appendChild(createNoResultsMessage());
-        loadingIndicator.style.display = 'none';
-        return;
-      }
-
-      const visibleItems = allItems.slice(0, page * ITEMS_PER_PAGE);
-      const newsBlock = buildBlock('news-cards', visibleItems);
-      
-      newsContentContainer.append(newsBlock, loadingIndicator);
-
-      
-      if (allItems.length > page * ITEMS_PER_PAGE) {
-        const moreButton = document.createElement('button');
-        moreButton.className = 'news-more-button';
-        moreButton.textContent = 'More News';
-        moreButton.addEventListener('click', () => {
-          currentPage++;
-          showItems(currentPage);
-        });
-        
-        loadingIndicator.style.display = 'none';
-        newsContentContainer.append(moreButton);
-      } else {
-        loadingIndicator.style.display = 'none';
-      }
-    };
-
-    /**
-     * Handles filter changes
-     * @param {string} filterValue - Selected filter value
-     */
-    const handleFilterChange = async (filterValue) => {
-      currentPage = 1;
-      loadingIndicator.style.display = 'block';
-      allItems = await fetchNewsItems(filterValue);
-      showItems(currentPage);
-    };
-
-    // Setup structure
-    contentSection.append(newsContentContainer);
-    block.append(createFilterSection(handleFilterChange), createDivider(),contentSection);
-    newsContentContainer.append(loadingIndicator);
-
-    // Initial load
-    const initialFilter = getFilterFromURL();
-    allItems = await fetchNewsItems(initialFilter);
-    if (allItems.length) {
-      showItems(currentPage);
-    }
-
+    // Load block
+    await loadBlock(newsSection);
   } catch (error) {
-    console.error('Error in news block decoration:', error);
+    console.error('Error in news landing decoration:', error);
   }
 }
